Split `Type` and `TypeSet` into separate files.
